修改taskcontroller中的获取参数

+++ src/main/java/com/bupt/battery/form/TaskQueryForm.java
@@ -25,5 +25,5 @@ public class TaskQueryForm implements Serializable {
     private String vehicleId;
     private List<ParamAO> paramAOList;
     private String shopId;
-    private Integer taskTypeId;
+    private Long taskTypeId;
 }
