+++ src/main/java/com/bupt/battery/service/ITaskDOService.java
@@ -8,5 +8,5 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface ITaskDOService extends BaseService<TaskDO,Long> {
-    Page<TaskDO> findTaskListByPage(Integer pageNum,Integer pageSize,String taskName, Date startTime,Date endTime, Integer taskTypeId);
+    Page<TaskDO> findTaskListByPage(Integer pageNum,Integer pageSize,String taskName, Date startTime,Date endTime, Long taskTypeId);
 }
